+++ CHANGELOG.md
@@ -4,6 +4,7 @@
   - fc_network support for API300
   - fcoe_network support for API300
   - network_set support for API300
+  - fabric new resource support for API300 Synergy
 
 ### 1.1.0
   - Add support for client ENV variables